# docs/swagger/swagger.yaml
...
info:
  contact:
    name: Metraly
  description: Team Engineering Metrics API
  title: Metraly API
  version: "1.0"
paths:
...
  /api/v1/role/{role}:
    get:
      consumes:
      - application/json
      description: Returns dashboard data for a specific role (engineer, lead, manager)
      parameters:
      - description: Role name
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get role-specific dashboard
      tags:
      - role
swagger: "2.0"